store shorthand flag as a single UTF-8 character
Currently the shorthand for a flag is stored as a string, which does not
make a lot of sense considering that there is a strict requirement for
the shorthand flag to only be one character long.

This can be better represented by storing the shorthand as a rune, which is
a datatype designed to store exactly one UTF-8 character. Actually, until
now the requirement was for the shorthand to be one *ASCII* character long.
This was enforced by panicking at a few points when len(shorthand) > 1.
Obviously, this is not ideal – think of a flag like 'ö' or '世'.
This is also revised with this patch, since we now actually check for
the *UTF-8* length of the string, and only panic when it contains too
many runes.

I intentionally did not change the public API of the package, though
this would obviously be even better, since some of the panics (like, in
ShorthandLookup) could be avoided by passing a rune instead of a string
in the first place.

Merges https://github.com/spf13/pflag/pull/264