Standardize flag API
There are many functions for each flag type, all with slight
differences compared to the other functions for the same flag type.
This commit changes that so there is only two functions per flag
type. The new functions ("<Type>Var", and "<Type>", e.g. `BoolVar()`,
and `Bool()`) have a single difference, one (`<Type>Var()`) which allows you to
pass in a pointer to which the value will be assigned, and the
other (`<Type>()`) does not. Additional options can be specified using
`Opt*` function calls, e.g. `OptShorthand`, or `OptShorthandDeprecated`,
etc. Here is an example:

```go
Bool("flag-name", false, "flag usage", OptShorthandOnly(true), OptShorthand('1'))
```

The `Opt*()` functions have implemented to avoid redudancy. E.g. boolean
options do not have a value, and are just assumed to be true. The option
should be removed to set to false. All options can be appended to both
the `<Type>Var` and `<Type>`, and multiple can be appended at the same
time.

The global flag functions have also been replaced with a single `Var()`
function with the same API as the `<Type>Var` function.

This is helpful as this allows customising a flag without having to
define it, then using the `Lookup()` function, and finally making the
relevant changes. Simplifying these function calls significantly reduces
the code base and makes it easier to understand.